requirements.txt
shiny
tensorflow-cpu
numpy
Pillow
pandas